[Refactor]既存コンポーネントを修正(molecules群)
